Fullscreen chart shared axis toggle
